url_shortner/requirements.txt
alembic==1.13.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
